Move monitoring bootstrap logic to monitoring/init.js
