+++ src/index.js
@@ -1,14 +1,4 @@
 import 'babel-polyfill';
-import merge from 'webpack-merge';
-import config from './config/config';
-import Monitor from './monitoring/Monitor';
-import * as eventListener from './monitoring/eventListener';
+import init from './monitoring/init';
 
-eventListener.listen();
-
-config.services.forEach(service => {
-  let serviceConfig = merge(config.monitoring, service);
-  let monitor = new Monitor(serviceConfig);
-
-  monitor.start();
-});
+init();
